+++ examples/pydantic_recursive.py
@@ -2,9 +2,9 @@
 This example demonstrates pydantic serialisation of a recursively cycled model.
 """
 from tortoise import Tortoise, fields, run_async
+from tortoise.contrib.pydantic import pydantic_model_creator
 from tortoise.exceptions import NoValuesFetched
 from tortoise.models import Model
-from tortoise.pydantic import pydantic_model_creator
 
 
 class Employee(Model):
