+++ tests/contrib/test_pydantic.py
@@ -1,6 +1,6 @@
 from tests.testmodels import Address, Employee, Event, Reporter, Team, Tournament
 from tortoise.contrib import test
-from tortoise.pydantic import pydantic_model_creator, pydantic_queryset_creator
+from tortoise.contrib.pydantic import pydantic_model_creator, pydantic_queryset_creator
 
 
 class TestPydantic(test.TestCase):
